+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 UNRELEASED
 ----------
 
+* [ [#2037](https://github.com/digitalfabrik/integreat-cms/issues/2037) ] Add organization and barrier-free fields to POI
+
 
 2023.1.2
 --------